requirement.txt
scanpy
pandas
anndata
scikit-learn
pytorch-metric-learning